SQLite.swift
I have used this to access the SQLite database. It is a separate Github project that is being updated. This is an earlier copy I dowloaded that worked at the time but is probably not updated. If you want the latest one, you might want to connect directly to that project.